requirements.txt
pandas==2.0.3
Pillow==8.2.0
Pillow==10.1.0
plotly==5.1.0
streamlit==1.25.0
streamlit_extras==0.3.4
